+++ regression-tests/long_double/init_3.c
@@ -1,6 +1,6 @@
 int main() {
-  double a = 4.24242f;
-  assert(a == 4.24241f);
+  long double a = 4.24242l;
+  assert(a == 4.24241l);
   return 0;
 
 }
