+++ regression-tests/long_double/int_plus_float.c
@@ -1,7 +1,7 @@
 int main ()
 {
-  double a = 2.0f;
+  long double a = 2.0l;
   int b = 1;
-  assert(a+b == 3.0f);
+  assert(a+b == 3.0l);
   return 0;
 }
